Add optional delay before showing popover
Add a delay before they popover is shown. The use case is attaching a
popover to buttons/links on hover. I'd like the user to be able to click
the element without necessarily seeing the popover but if they hesitate
then they see a help message in a popover (something more elaborate than
a regular tooltip).